Refactored UI component codes
