+++ app/src/main/java/com/kyhsgeekcode/disassembler/Utils/Util.java
@@ -9,6 +9,7 @@ import android.provider.MediaStore;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -60,4 +61,12 @@ public class Util {
         return filePath;
     }
 
+    //https://stackoverflow.com/a/6425744/8614565
+    public static void deleteRecursive(File fileOrDirectory) {
+        if (fileOrDirectory.isDirectory())
+            for (File child : fileOrDirectory.listFiles())
+                deleteRecursive(child);
+
+        fileOrDirectory.delete();
+    }
 }
